Update Dockerfile
Hi, this is Mizzlr. I have updated your dockerfile to reduce the image size. The size of the original dockerfile is about 222MB and size of the modified dockerfile 178 MB. I have docker built and tested the modified image and it works fine. Please accept my proposal for file change. Thank you :+1 :)